Finalizes FixMe container
Specify Action (Solution, Start, Stop etc)
Exercise code (eg. DEVWF-T004)
Source repo (default: https://github.com/XpiritBV/CodeToCloud-Workshop)

And a pull request will be created against the defaut compare branch in the target repo (configure in settings.json)

+++ Containers/ghws-fix/fixme.ps1
@@ -127,27 +127,55 @@ function Run {
     $settings = Get-GlobalSettings
 
     $token = $settings.GithubToken
-    $sourceRepo = $env.SOURCEREPO
+    $sourceRepo = $env:SOURCEREPO
     $targetRepo = $settings.TargetRepo
+
+    $action = $env:ACTION
+    $exercise = $env:EXERCISE
     
     $sourceRepoFolder = Join-Path -path $pwd "source"
     $targetRepoFolder = Join-Path -path $pwd "target"
 
     $token | gh auth login --with-token 
-    Remove-Item $sourceRepoFolder -Recurse -Force
-    Remove-Item $targetRepoFolder -Recurse -Force
 
-    $markDown = Get-PullRequestProperties (Join-Path -path $sourceRepoFolder "pr.md")
-
-    $branch = $markDown.Properties.Branch
+    if (Test-Path -PathType Container $sourceRepoFolder) {
+        Remove-Item $sourceRepoFolder -Recurse -Force
+    }
+    if (Test-Path -PathType Container $targetRepoFolder) {
+        Remove-Item $targetRepoFolder -Recurse -Force
+    }
 
     Invoke-GitHub repo clone $sourceRepo $sourceRepoFolder
+
+    # Find the right folder based on **/action/exercise/
+    $markdownPathRoot = Get-ChildItem -Path $sourceRepoFolder -Include pr.md -File -Recurse | 
+        where-object { $_.DirectoryName -like "*`\$action`\$exercise*" } | 
+        select-object -ExpandProperty "DirectoryName" { $_.DirectoryNamey } 
+
+    $markdownPath = Join-Path -path $markdownPathRoot "pr.md" 
+    if (-not (Test-Path -Path $markdownPath)) {
+        throw "Missing markdown file, expected here: $markdownPath"
+    }
+
+    $markDown = Get-PullRequestProperties $markdownPath
+    $branch = $markDown.Properties.Branch
+
     Invoke-GitHub repo clone $targetRepo $targetRepoFolder
 
     pushd $targetRepoFolder
 
+    #To reset the target environment, run below line
     #Remove-Item * -Exclude ".git" -Recurse
-    Copy-Item (Join-Path -path $sourceRepoFolder "./contents/*") "."
+    
+    $solutionContentsPath = Join-Path -path $markdownPathRoot "contents"
+    if (-not (Test-Path -PathType Container $solutionContentsPath)) {
+        throw "Missing solution content, expected here: $solutionContentsPath"
+    }
+    Copy-Item (Join-Path -path $solutionContentsPath "*") $targetRepoFolder -Recurse
+
+    Invoke-Git config user.name "$($markDown.Properties.Committer)"
+    Invoke-Git config user.email "$($markDown.Properties.CommitterEmail)"
+
     Invoke-Git checkout -B $branch 
     Invoke-Git add . 
     Invoke-Git commit -m $markDown.Properties.CommitMessage
